fix style to following pep8

+++ mob2exec/bin/sim_submission.py
@@ -43,19 +43,19 @@ from mobyle.common.job import Status
 from mobyle.common.job import ClJob
 
 
-cmdlines={
-        1:{ 'name':'blast', 
+cmdlines = {
+        1 : { 'name':'blast', 
              'cmd_line': 'blastall -p blastp -d uniprot_sprot -i abcd2_mouse.fa -e 0.1 -o blast2.txt',
              'inputs':[( 'abcd2_mouse.fa', """>ABCD2_MOUSE RecName: Full=ATP-binding cassette sub-family D member 2; AltName: Full=Adrenoleukodystrophy-related protein;
 MIHMLNAAAYRVKWTRSGAAKRAACLVAAAYALKTLYPIIGKRLKQPGHRKAKAEAYSPAENREILHCTEIICKKPAPGL
 NAAFFKQLLELRKILFPKLVTTETGWLCLHSVALISRTFLSIYVAGLDGKIVKSIVEKKPRTFIIKLIKWLMIAIPATFV"""
               )]
              },
-        2:{ 'name': 'golden', 
+        2 : { 'name': 'golden', 
             'cmd_line': 'golden uniprot_sprot:abcd2_rat',
             'inputs':[]
             },
-        3:{ 'name': 'clustalw',
+        3 : { 'name': 'clustalw',
             'cmd_line':'clustalw -align -infile=abcd.fasta',
             'inputs' : [( 'abcd.fasta' , """>1pgb
 MTYKLILNGKTLKGEAVDAATAEKVFKQYANDNGVDGEWTYTKTFTVTE
@@ -75,10 +75,11 @@ def put_new_job_in_db(name, cmd_line, project ):
     job = connection.ClJob()
     job.project = project.id 
     job.name = name
-    job.status = Status(Status.BUILDING)
+    job.status = Status(Status.TO_BE_BUILT)
     job.owner = {'id': project.id, 'klass': 'Project'}
     job.cmd_line = cmd_line
     job.save()
+    print("put new job {0} in db".format(job.id))
 
 def clean_db():
     old_jobs = connection.Job.find({})
@@ -96,7 +97,8 @@ def clean_db():
         
 def clean_dirs():
     projects_dir = os.path.join(os.path.dirname(config.get("mob2exec","pid_file")), 'projects')
-    shutil.rmtree(projects_dir)
+    if os.path.exists(projects_dir):
+        shutil.rmtree(projects_dir)
     
 def create_user(name):
     user = connection.User()
@@ -109,8 +111,12 @@ def create_project(user, name):
     project['owner'] = user['_id']
     project['name'] = name
     project.save()
-    project_directory = os.path.join(os.path.dirname(config.get("mob2exec","pid_file")), 'projects', str(project.id))   
+    project_directory = os.path.join(os.path.dirname(config.get("mob2exec","pid_file")), 
+                                     'projects', 
+                                     str(project.id))   
     os.makedirs(project_directory, 0755) #create parent directory
+    project.dir = project_directory
+    project.save()
     return project
 
 if __name__ == '__main__':
@@ -121,13 +127,12 @@ if __name__ == '__main__':
     for name in ('Filochard', 'Ribouldingue', 'Croquignol'):
         pieds_nickeles[name] = create_user(name)
     
-    for name in ('organisateurs de voyage' , 'dans le cambouis' , "l'Opération congélation"):
+    for name in ('organisateurs de voyage', 'dans le cambouis', "l'Opération congélation"):
         projects[name] = create_project(random.choice(pieds_nickeles.values()), name)
     
     for i in range(0, 500):
         time.sleep(random.randint(0, 5))
-        j = cmdlines[ random.randint(1,3)]
-        print( "put new job %s in db" % j['name'])
-        put_new_job_in_db( j['name'], j['cmd_line'], random.choice(projects.values()))
+        j = cmdlines[ random.randint(1, 3)]
+        put_new_job_in_db(j['name'], j['cmd_line'], random.choice(projects.values()))
            
-    print( "no more job to add in DB")
+    print("no more job to add in DB")
